+++ src/lib.rs
@@ -1,95 +1,5 @@
-use std::sync::{mpsc, Arc, Mutex};
-use std::thread;
-
 pub mod handlers;
 pub mod link_handler;
 pub mod router;
 pub mod server;
-
-pub struct ThreadPool {
-    workers: Vec<Worker>,
-    sender: mpsc::Sender<Message>,
-}
-
-type Job = Box<dyn FnOnce() + Send + 'static>;
-
-enum Message {
-    NewJob(Job),
-    Terminate,
-}
-
-impl ThreadPool {
-    /// Create a ThreadPool with specific size
-    ///
-    /// # Panics
-    /// Function will panic if size is less or equal to 0
-    pub fn new(size: usize) -> ThreadPool {
-        assert!(size > 0);
-
-        let (sender, receiver) = mpsc::channel();
-
-        let receiver = Arc::new(Mutex::new(receiver));
-
-        let mut workers = Vec::with_capacity(size);
-        for id in 0..size {
-            workers.push(Worker::new(id, Arc::clone(&receiver)));
-        }
-
-        ThreadPool { workers, sender }
-    }
-
-    pub fn execute<F>(&self, f: F)
-    where
-        F: FnOnce() + Send + 'static,
-    {
-        let job = Box::new(f);
-        self.sender.send(Message::NewJob(job)).unwrap();
-    }
-}
-
-impl Drop for ThreadPool {
-    fn drop(&mut self) {
-        println!("sending terminate message to all workers ...");
-        for _ in &self.workers {
-            self.sender.send(Message::Terminate).unwrap();
-        }
-
-        // workers received the terminate message, now wait for them to finish
-        // their current jobs, and break; from their loops
-
-        for worker in &mut self.workers {
-            println!("shutting down worker {}", worker.id);
-
-            if let Some(thread) = worker.thread.take() {
-                thread.join().unwrap();
-            }
-        }
-    }
-}
-
-struct Worker {
-    id: usize,
-    thread: Option<thread::JoinHandle<()>>,
-}
-
-impl Worker {
-    fn new(id: usize, receiver: Arc<Mutex<mpsc::Receiver<Message>>>) -> Worker {
-        let thread = thread::spawn(move || loop {
-            let message = receiver.lock().unwrap().recv().unwrap();
-            match message {
-                Message::NewJob(job) => {
-                    println!("worker {} got a new job, executing ...", id);
-                    job();
-                }
-                Message::Terminate => {
-                    println!("worker {} received termination message", id);
-                    break;
-                }
-            }
-        });
-        Worker {
-            id,
-            thread: Some(thread),
-        }
-    }
-}
+pub mod thread_pool;
